backend/pyproject.toml
[project]
name = "pyaglogen3d-backend"
version = "0.1.0"
description = "Backend for pyAgloGen3D - 3D Agglomerate Simulation and Fractal Analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Juanjo" }
]

dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "drf-nested-routers>=0.93",
    "django-cors-headers>=4.3",
    "psycopg[binary]>=3.1",
    "celery[redis]>=5.3",
    "redis>=5.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "scikit-image>=0.22",
    "pillow>=10.0",
    "matplotlib>=3.8",
    "python-decouple>=3.8",
    "gunicorn>=21.0",
    "whitenoise>=6.6",
    "redis-cli>=1.0.1",
    # Authentication
    "djangorestframework-simplejwt>=5.3",
    "dj-rest-auth>=7.0",
    "django-allauth[socialaccount]>=65.0",
    # AI Assistant dependencies
    "anthropic>=0.40",
    "openai>=1.50",
    "cryptography>=43.0",
    "jsonschema>=4.20.0",
    # RAG dependencies
    "pgvector>=0.3.0",
    "langchain-text-splitters>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.7",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.1",
    "mypy>=1.7",
    "django-stubs>=4.2",
    "ipython>=8.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "apps"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "DJ"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
plugins = ["mypy_django_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
addopts = "-v --tb=short"
